ai_orchestrator/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-orchestrator"
version = "0.1.0"
description = "AI orchestrator service for date planning chatbot"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "grpcio",
    "grpcio-tools", 
    "protobuf",
    "uvloop>=0.20.0; platform_system!='Windows'",
    "pydantic>=2.7",
    "httpx>=0.27",
    "python-dotenv",
    "openai>=1.40.0",
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "isort",
    "mypy",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["ai_orchestrator*"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
